# non-allocating version of AllPerms