benchmark:
```julia
function generate(itr)
    k = 0
    for p in itr
        @assert p isa perm
        k += p[1]
    end
    return k
end
```
## Old:
```julia
julia> @time generate(elements(PermGroup(10));
  3.818291 seconds (36.29 M allocations: 1.352 GiB, 5.59% gc time)

julia> @time generate(UInt(10));
  3.478949 seconds (36.29 M allocations: 1.352 GiB, 6.14% gc time)
```
## New:
```julia
julia> @time K1 = generate(elements(PermGroup(10)));
  0.307158 seconds (7.26 M allocations: 664.454 MiB, 22.96% gc time)

julia> @time K2 = generate(Generic.elements!(PermGroup(10)));
  0.035850 seconds (12 allocations: 640 bytes)

julia> K1 == K2
true
```
In `generate(elements(PermGroup(10)))`:
* `p = randperm(10); perm(copy(p), false)` allocates `192` bytes and `192*factorial(10)/1024/1024 = 664.453125` (MiB of memory). I don't think You can get any better with allocating Heaps;
* there are two allocations per permutation (`perm` and `copy`?)

This puts us in the league of `C` when comparing to https://en.wikipedia.org/wiki/Heap's_algorithm#cite_note-3: For a simple loop crawling over all perms:
```julia
function allperms(N)
    elts = collect(1:N)
    c = ones(Int, N)
#     doit(elts)
    n = 1
    k = 0
    cc = 1
    while n <= length(elts)
      if c[n] < n
         k = ifelse(isodd(n), 1, c[n])
         elts[k], elts[n] = elts[n], elts[k]
         c[n] += 1
         n = 1
            cc += 1
#          doit(elts)
      else
        c[n] = 1
        n += 1
      end
   end
   return cc
end
@assert allperms(5) == factorial(5)
@time allperms(12)
@time generate(N, Generic.elements!(PermGroup(12)))
```
> `  3.373536 seconds (8 allocations: 560 bytes)`
> `  4.107824 seconds (11 allocations: 640 bytes)`