@wbhart 2018-05-11
The new timings look great!